requirements.txt
Django~=3.2.10
django-ndarray==0.0.3
django-environ==0.8.1
imageio==2.14.1
matplotlib==3.5.1
numpy==1.21.4
Pillow==9.0.0